feat: implement basic search field
The search field appears when `searchFilters` are specified in the instance configuration (see README.md).

It's supposes to work whith any number of filters but currently leads to odd results at the moment. This will be tricky to get it working with 0 config...

Simulating a click on a feature was a pain ! This was in fact highly coupled to deck-gl and the interface.

There is still an odd bug :
If you never hover the map with the mouse cursor the popover won't open when you click on a feature in the search results...

chore: remove versioned test files

chore: add test files to git ignore list

debugging

fix: fix search using standard select

This could use a debounce tho'